Used custom version of go-client for seed only cluster.
The operator when operating on the cluster should use the pod ips
only. This is because the access and alt-access addresses might not be
accessible for the operator.

TODO: Added a custom policy to the go client. If it passes the review
and is a reasonable request will get this new policy merged into
Aerospike go client.