fix: make role create code stop always returning err in isCreate (#411)
There was a bug in aerospikeRoleCreateUpdate.execute where isCreate,
reached when err from QueryRole contains "Invalid role", shadows that
same err variable.

The err variable is shadowed in the "if err :=" block, goes out of
scope, and the previous err ("Invalid role") is returned instead of nil.
Fix this by renaming the error to avoid shadowing.